Add and publish timestamps for RFID and plug_state
Timestamp plug_time for plug_state was already defined in
ChargepointData.set and used in ChargepointRfidMixin, but I could not
find any codepart, setting it. Also could not see it being set on
productive System. It is neccessary for correctly assigning RFID on a
Duo to the LP plugged in last.
Timestamp for RFID Scan is neccessary for handling RFIDs with a timeout.